+++ test/src/ExtraBuildResourcesTest.ts
@@ -119,7 +119,7 @@ test.ifAll.ifNotWindows(
             publish: null,
             // https://github.com/electron-userland/electron-builder/issues/1355
             linux: {
-              target: ["AppImage", "deb", "rpm"],
+              target: ["AppImage", "deb", "rpm", "pacman"],
             },
             compression: "store",
           },
@@ -172,6 +172,18 @@ test.ifNotWindows(
     targets: linuxDirTarget,
     config: {
       electronDist: getElectronCacheDir(),
+    },
+  })
+)
+
+test.ifNotWindows(
+  "electronDist as callback function for path to local folder with electron builds zipped ",
+  app({
+    targets: linuxDirTarget,
+    config: {
+      electronDist: (_context) => {
+        return Promise.resolve(getElectronCacheDir())
+      }
     },
   })
 )