+++ test/src/globTest.ts
@@ -141,7 +141,7 @@ test.ifAll.ifDevOrLinuxCi("ignore node_modules", () => {
           data.dependencies = {
             "ci-info": "2.0.0",
             // this contains string-width-cjs 4.2.3
-            "@isaacs/cliui":"8.0.2"
+            "@isaacs/cliui": "8.0.2",
           }
         }),
       packed: context => {